update queing
git-svn-id: https://build.archivesportaleurope.net/svn/apex/trunk@161 9fd842fd-667d-44da-918d-ff91420b3ae1

+++ Dashboard/src/main/java/eu/apenet/dashboard/services/ead/EadService.java
@@ -1,6 +1,7 @@
 package eu.apenet.dashboard.services.ead;
 
 import java.util.Date;
+import java.util.List;
 
 import org.apache.log4j.Logger;
 
@@ -8,6 +9,7 @@ import eu.apenet.commons.types.XmlType;
 import eu.apenet.commons.utils.APEnetUtilities;
 import eu.apenet.dashboard.security.SecurityContext;
 import eu.apenet.persistence.dao.EadDAO;
+import eu.apenet.persistence.dao.EadSearchOptions;
 import eu.apenet.persistence.dao.QueueItemDAO;
 import eu.apenet.persistence.factory.DAOFactory;
 import eu.apenet.persistence.vo.Ead;
@@ -89,7 +91,7 @@ public class EadService {
 				}
 			} else { // Add the file to the indexing queue
 				try {
-					addToQueue(ead, QueueAction.INDEX);
+					addToQueue(ead, QueueAction.PUBLISH);
 					processed = true;
 				} catch (Exception e) {
 					LOGGER.error("Unable to put: " + ead + " in a queue: " + e.getMessage(), e);
@@ -99,21 +101,41 @@ public class EadService {
 		return processed;
 	}
 
+	public static boolean convertValidatePublish(XmlType xmlType, Integer id) {
+		EadDAO eadDAO = DAOFactory.instance().getEadDAO();
+		Ead ead = eadDAO.findById(id, xmlType.getClazz());
+		SecurityContext.get().checkAuthorized(ead);
+		if (!ead.isSearchable()) {
+			addToQueue(ead, QueueAction.CONVERT_VALIDATE_PUBLISH);
+		}
+		return true;
+	}
+
 	public static boolean unpublish(XmlType xmlType, Integer id) {
 		EadDAO eadDAO = DAOFactory.instance().getEadDAO();
 		Ead ead = eadDAO.findById(id, xmlType.getClazz());
 		SecurityContext.get().checkAuthorized(ead);
 		boolean processed = false;
 		if (ead.isSearchable()) {
-			try {
-				JpaUtil.beginDatabaseTransaction();
-				new UnpublishTask().execute(ead);
-				JpaUtil.commitDatabaseTransaction();
-				processed = true;
-			} catch (Exception e) {
-				JpaUtil.rollbackDatabaseTransaction();
-				LOGGER.error("Unable to unpublish: " + ead + " : " + e.getMessage(), e);
+			if (APEnetUtilities.getDashboardConfig().isDirectIndexing()) {
+				try {
+					JpaUtil.beginDatabaseTransaction();
+					new UnpublishTask().execute(ead);
+					JpaUtil.commitDatabaseTransaction();
+					processed = true;
+				} catch (Exception e) {
+					JpaUtil.rollbackDatabaseTransaction();
+					LOGGER.error("Unable to unpublish: " + ead + " : " + e.getMessage(), e);
+				}
+			} else { // Add the file to the indexing queue
+				try {
+					addToQueue(ead, QueueAction.UNPUBLISH);
+					processed = true;
+				} catch (Exception e) {
+					LOGGER.error("Unable to put: " + ead + " in a queue: " + e.getMessage(), e);
+				}
 			}
+
 		}
 		return processed;
 	}
@@ -123,17 +145,25 @@ public class EadService {
 		Ead ead = eadDAO.findById(id, xmlType.getClazz());
 		SecurityContext.get().checkAuthorized(ead);
 		boolean processed = false;
-		try {
-			JpaUtil.beginDatabaseTransaction();
-			new UnpublishTask().execute(ead);
-			new DeleteTask().execute(ead);
-			JpaUtil.commitDatabaseTransaction();
-			processed = true;
-		} catch (Exception e) {
-			JpaUtil.rollbackDatabaseTransaction();
-			LOGGER.error("Unable to delete: " + ead + " : " + e.getMessage(), e);
+		if (APEnetUtilities.getDashboardConfig().isDirectIndexing()) {
+			try {
+				JpaUtil.beginDatabaseTransaction();
+				new UnpublishTask().execute(ead);
+				new DeleteTask().execute(ead);
+				JpaUtil.commitDatabaseTransaction();
+				processed = true;
+			} catch (Exception e) {
+				JpaUtil.rollbackDatabaseTransaction();
+				LOGGER.error("Unable to delete: " + ead + " : " + e.getMessage(), e);
+			}
+		} else { // Add the file to the indexing queue
+			try {
+				addToQueue(ead, QueueAction.DELETE);
+				processed = true;
+			} catch (Exception e) {
+				LOGGER.error("Unable to put: " + ead + " in a queue: " + e.getMessage(), e);
+			}
 		}
-
 		return processed;
 	}
 
@@ -146,18 +176,29 @@ public class EadService {
 		ead.setQueuing(QueuingState.BUSY);
 		eadDAO.store(ead);
 		try {
-			if (queueItem.getAction().isConvertAction() && !ead.isConverted()) {
-				new ConvertTask().execute(ead);
+			if (queueItem.getAction().isDeleteAction()) {
+				JpaUtil.beginDatabaseTransaction();
+				new UnpublishTask().execute(ead);
+				new DeleteTask().execute(ead);
+				queueItemDAO.deleteSimple(queueItem);
+				JpaUtil.commitDatabaseTransaction();
+			} else {
+				if (queueItem.getAction().isConvertAction() && !ead.isConverted()) {
+					new ConvertTask().execute(ead);
+				}
+				if (queueItem.getAction().isValidateAction() && ValidatedState.NOT_VALIDATED.equals(ead.getValidated())) {
+					new ValidateTask().execute(ead);
+				}
+				if (queueItem.getAction().isPublishAction() && ValidatedState.VALIDATED.equals(ead.getValidated())
+						&& !ead.isSearchable()) {
+					new PublishTask().execute(ead);
+				}
+				ead.setQueuing(QueuingState.NO);
+				eadDAO.store(ead);
+				queueItemDAO.delete(queueItem);
 			}
-			if (queueItem.getAction().isValidateAction()  && ValidatedState.NOT_VALIDATED.equals(ead.getValidated())) {
-				new ValidateTask().execute(ead);
-			}
-			if (queueItem.getAction().isIndexAction() && ValidatedState.VALIDATED.equals(ead.getValidated()) && !ead.isSearchable()) {
-				new PublishTask().execute(ead);
-			}
-			ead.setQueuing(QueuingState.NO);
-			eadDAO.store(ead);
-			queueItemDAO.delete(queueItem);
+
+
 			processed = true;
 		} catch (Exception e) {
 			String err = "eadid: " + ead.getEadid() + " - id: " + ead.getId() + " - type: " + xmlType.getName();
@@ -190,11 +231,66 @@ public class EadService {
 			queueItem.setSourceGuide((SourceGuide) ead);
 			priority += 50;
 		}
-		if (queueAction.isConvertAction() || queueAction.isValidateAction() || queueAction.isIndexAction()) {
+		if (queueAction.isConvertAction() || queueAction.isValidateAction() || queueAction.isPublishAction()) {
 			priority += 25;
 		}
 		queueItem.setPriority(priority);
 		indexqueueDao.store(queueItem);
 	}
 
+	public static void addBatchToQueue(Integer aiId, XmlType xmlType, QueueAction queueAction) {
+		SecurityContext.get().checkAuthorized(aiId);
+		QueueItemDAO indexqueueDao = DAOFactory.instance().getQueueItemDAO();
+		EadDAO eadDAO = DAOFactory.instance().getEadDAO();
+		EadSearchOptions eadSearchOptions = new EadSearchOptions();
+		eadSearchOptions.setPageSize(1000);
+		if (QueueAction.CONVERT.equals(queueAction)) {
+			eadSearchOptions.setConverted(false);
+		} else if (QueueAction.VALIDATE.equals(queueAction)) {
+			eadSearchOptions.setConverted(true);
+			eadSearchOptions.setValidated(ValidatedState.NOT_VALIDATED);
+		} else if (QueueAction.PUBLISH.equals(queueAction)) {
+			eadSearchOptions.setConverted(true);
+			eadSearchOptions.setValidated(ValidatedState.VALIDATED);
+			eadSearchOptions.setSearchable(false);
+		} else if (QueueAction.CONVERT_VALIDATE_PUBLISH.equals(queueAction)) {
+			eadSearchOptions.setSearchable(false);
+		} else if (QueueAction.UNPUBLISH.equals(queueAction)) {
+			eadSearchOptions.setSearchable(true);
+		}
+		eadSearchOptions.setEadClazz(xmlType.getClazz());
+		eadSearchOptions.setArchivalInstitionId(aiId);
+		JpaUtil.beginDatabaseTransaction();
+		List<Ead> eads = eadDAO.getEads(eadSearchOptions);
+		int size = 0;
+		while ((size = eads.size()) > 0) {
+			Ead ead = eads.get(size - 1);
+			QueueItem queueItem = new QueueItem();
+			queueItem.setQueueDate(new Date());
+			queueItem.setAction(queueAction);
+			ead.setQueuing(QueuingState.READY);
+			eadDAO.updateSimple(ead);
+			int priority = 0;
+			if (ead instanceof FindingAid) {
+				queueItem.setFindingAid((FindingAid) ead);
+			} else if (ead instanceof HoldingsGuide) {
+				queueItem.setHoldingsGuide((HoldingsGuide) ead);
+				priority += 100;
+			} else if (ead instanceof SourceGuide) {
+				queueItem.setSourceGuide((SourceGuide) ead);
+				priority += 50;
+			}
+			if (queueAction.isConvertAction() || queueAction.isValidateAction() || queueAction.isPublishAction()) {
+				priority += 25;
+			}
+			if (queueAction.isDeleteAction() || queueAction.isUnpublishAction()) {
+				priority += 1000;
+			}
+			queueItem.setPriority(priority);
+			eads.remove(size - 1);
+			indexqueueDao.updateSimple(queueItem);
+		}
+		JpaUtil.commitDatabaseTransaction();
+
+	}
 }
